+++ src/pymmcore_widgets/useq_widgets/__init__.py
@@ -1,5 +1,6 @@
 """Widgets for the useq-schema data model."""
 
+from ._channels import ChannelTable
 from ._column_info import (
     BoolColumn,
     FloatColumn,
@@ -8,11 +9,10 @@ from ._column_info import (
     TimeDeltaColumn,
 )
 from ._data_table import DataTable, DataTableWidget
-from .channels import ChannelTable
-from .mda_sequence import MDASequenceWidget
-from .positions import PositionTable
-from .time import TimeTable
-from .z import ZPlanWidget
+from ._mda_sequence import MDASequenceWidget
+from ._positions import PositionTable
+from ._time import TimeTable
+from ._z import ZPlanWidget
 
 __all__ = [
     "PositionTable",
