+++ src/pymmcore_widgets/useq_widgets/_channels.py
